Move xaml to /ui/dialog
