+++ Galaxy Buds Client/ui/dialog/BudsPopup.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="Galaxy_Buds_Client.BudsPopup"
+﻿<Window x:Class="Galaxy_Buds_Client.ui.dialog.BudsPopup"
 		xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 		xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 		xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
