+++ Galaxy Buds Client/ui/dialog/BudsPopup.xaml.cs
@@ -1,22 +1,13 @@
-﻿using Galaxy_Buds_Client.message;
-using Galaxy_Buds_Client.model.Constants;
-using Galaxy_Buds_Client.transition;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System;
 using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
 using System.Windows.Media.Animation;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
+using Galaxy_Buds_Client.message;
+using Galaxy_Buds_Client.model.Constants;
+using Galaxy_Buds_Client.transition;
 
-namespace Galaxy_Buds_Client {
+namespace Galaxy_Buds_Client.ui.dialog {
     /// <summary>
     /// Interaction logic for BudsPopup.xaml
     /// </summary>
